+++ src/Common/index.js
@@ -1,0 +1,11 @@
+import Buttons from './Buttons'
+import CollapsibleList from './CollapsibleList'
+import Header from './Header'
+import InputField from './InputField'
+
+export {
+	Buttons,
+	CollapsibleList,
+	Header,
+	InputField
+}
