+++ src/index.js
@@ -1,0 +1,19 @@
+import Carousel from './Carousel';
+import Home from './Home';
+import Brochure from './Menu';
+import Faq from './Faq';
+import CutOffs from './Cutoffs';
+import Centers from './Centers';
+import Report from './Report';
+import MenuControls from './Menu/MenuControls';
+
+export {
+	Carousel,
+	Home,
+	Brochure,
+	Faq,
+	CutOffs,
+	Centers,
+	Report,
+	MenuControls
+}
